烂版信标测距效果

+++ spin27/src/common.h
@@ -5,8 +5,10 @@
 #include "MM32SPIN2xx_p.h"
 typedef struct
 {
-    uint16_t x;
-    uint16_t y;
+    // uint16_t x;
+    // uint16_t y;
+    int32_t x;
+    int32_t y;
 } point;
 typedef union
 {
